Added ParseXML class and ShotCounter class. All Room classes (and subclasses) now store coordinate data of their Scene card slots. ShotCounter objects also store their own separate coordinate data as well.
ParseXML currently only parses the board.xml file to construct a
Board object, complete with Rooms, neighbors (Room adjacency), on-Room roles,
etc. Next step is to add parsing of Scene cards as well.

Following that, is to make the UserView interactable with the primary
game loop to capture GUI / visual user input.